Add sprite that can be controlled by player
