+++ .gitignore
@@ -103,3 +103,6 @@ crashlytics-build.properties
 *.tga
 *.tif
 *.tiff
+
+# Files generated by JetBrains Rider
+.idea/
