+++ Assets/Scripts/BulletManagement/BulletPattern.cs
@@ -1,16 +1,19 @@
 ﻿using UnityEngine;
 
-/**
- * Defines bullet patterns for each bullet.
- * Requires a time t, representing how long a bullet has lived.
- */
-public class BulletPattern : MonoBehaviour
+namespace BulletManagement
 {
-    // Calculates translation when given time t of the lifetime
-    // of an object from float. bulletId is used if each bullet
-    // needs to have unique patterns.
-    public virtual Vector3 GetTranslation(float time, int bulletId)
+    /**
+    * Defines bullet patterns for each bullet.
+    * Requires a time t, representing how long a bullet has lived.
+    */
+    public class BulletPattern : MonoBehaviour
     {
-        return new Vector3(time, 0, 0);
+        // Calculates translation when given time t of the lifetime
+        // of an object from float. bulletId is used if each bullet
+        // needs to have unique patterns.
+        public virtual Vector3 GetTranslation(float time, int bulletId)
+        {
+            return new Vector3(time, 0, 0);
+        }
     }
 }
