+++ Assets/Scripts/BulletManagement/Patterns/OctoStraight.cs
@@ -1,21 +1,22 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
-/**
+namespace BulletManagement.Patterns
+{
+    /**
  * Emitts a pattern from the center in a straight line with 8 spokes.
  */
-public class OctoStraight : BulletPattern
-{
-    public override Vector3 GetTranslation(float time, int bulletId)
+    public class OctoStraight : BulletPattern
     {
-        float spokeCount = bulletId % 8;
+        public override Vector3 GetTranslation(float time, int bulletId)
+        {
+            float spokeCount = bulletId % 8;
 
-        float c = (Mathf.PI / 4) * spokeCount;
-        // Derivative functions with respect to time
-        float dx = Mathf.Cos(c);
-        float dy = Mathf.Sin(c);
+            float c = (Mathf.PI / 4) * spokeCount;
+            // Derivative functions with respect to time
+            float dx = Mathf.Cos(c);
+            float dy = Mathf.Sin(c);
 
-        return new Vector3(dx, dy, 0);
+            return new Vector3(dx, dy, 0);
+        }
     }
 }
