+++ Assets/Scripts/Level/Level.cs
@@ -1,71 +1,69 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
-/**
- * Logic associated with a Unity Scene.
- * Can handle movement of objects and instantiation of them
- * Generally you want anything that should be at the start of the game to be
- * placed in the scene first. For example the Camera which should already be in the Scene.
- */
-public class Level : MonoBehaviour
+namespace Level
 {
-    // Camera of the scene
-    Camera mainCamera;
-
-    public string mainLightName = "Main Light";
-    Light mainLight;
-
-    void InitCamera()
+    /**
+    * Logic associated with a Unity Scene.
+    * Can handle movement of objects and instantiation of them
+    * Generally you want anything that should be at the start of the game to be
+    * placed in the scene first. For example the Camera which should already be in the Scene.
+    */
+    public class Level : MonoBehaviour
     {
-        // Enable the main camera for usage
-        this.mainCamera.enabled = true;
-        this.mainCamera.transform.position = new Vector3(0, 0, -10);
+        // Camera of the scene
+        Camera mainCamera;
 
-        // Look towards negate Z axis
-        this.mainCamera.transform.forward = new Vector3(0, 0, 1);
-        this.mainCamera.rect = new Rect(0f, 0f, 1f, 1f);
+        public string mainLightName = "Main Light";
+        Light mainLight;
 
-        // Background color instead of default skyline
-        this.mainCamera.backgroundColor = Color.black;
-        this.mainCamera.clearFlags = CameraClearFlags.SolidColor;
-    }
+        private void InitCamera()
+        {
+            // Enable the main camera for usage
+            this.mainCamera.enabled = true;
+            this.mainCamera.transform.position = new Vector3(0, 0, -10);
 
-    void InitSun()
-    {
-        // Grab a directional light if possible
-        // otherwise create one
-        foreach (Light light in FindObjectsOfType<Light>()) {
-            if (light.type == LightType.Directional && light.name == this.mainLightName) {
-                this.mainLight = light;
-                break;
+            // Look towards negate Z axis
+            this.mainCamera.transform.forward = new Vector3(0, 0, 1);
+            this.mainCamera.rect = new Rect(0f, 0f, 1f, 1f);
+
+            // Background color instead of default skyline
+            this.mainCamera.backgroundColor = Color.black;
+            this.mainCamera.clearFlags = CameraClearFlags.SolidColor;
+        }
+
+        private void InitSun()
+        {
+            // Grab a directional light if possible
+            // otherwise create one
+            foreach (Light light in FindObjectsOfType<Light>()) 
+            {
+                if (light.type == LightType.Directional && light.name == this.mainLightName) {
+                    this.mainLight = light;
+                    break;
+                }
             }
+
+            if (this.mainLight == null) 
+            {
+                GameObject lightGameObject = new GameObject(this.mainLightName);
+                this.mainLight = lightGameObject.AddComponent<Light>();
+                this.mainLight.type = LightType.Directional;
+            }
+
+            // Face the light as the same direction as the camera
+            this.mainLight.transform.position = new Vector3(0, 0, 0);
+            this.mainLight.transform.forward = new Vector3(0, 0, 1);
         }
 
-        if (this.mainLight == null) {
-            GameObject lightGameObject = new GameObject(this.mainLightName);
-            this.mainLight = lightGameObject.AddComponent<Light>();
-            this.mainLight.type = LightType.Directional;
+        // Start is called before the first frame update
+        private void Start()
+        {
+            // initialize class variables
+            this.mainCamera = Camera.main;
+
+            // Initialize Camera
+            this.InitCamera();
+            this.InitSun();
         }
-
-        // Face the light as the same direction as the camera
-        this.mainLight.transform.position = new Vector3(0, 0, 0);
-        this.mainLight.transform.forward = new Vector3(0, 0, 1);
-    }
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        // initialize class variables
-        this.mainCamera = Camera.main;
-
-        // Initialize Camera
-        this.InitCamera();
-        this.InitSun();
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
     }
 }
