+++ Assets/Scripts/Movement/InputMove.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c98a1eebeb7f97f49b8a08a3e4c8f0c7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
